# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
security: []
paths:
  /users:
    get:
      operationId: listUsers
      summary: Get users list
      responses:
        '200':
          description: Users list successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    email:
                      type: string
                      format: email
                    name:
                      type: string
                    phone:
                      type: string
                  required:
                    - email
                    - name
        '401':
          description: Not authorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error or response contract violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Get current one user
      responses:
        '200':
          description: Current user successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    format: email
                  name:
                    type: string
                  phone:
                    type: string
                required:
                  - name
                  - email
        '401':
          description: Current user is not authorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error or response contract violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /products:
    get:
      operationId: listProducts
      summary: Get products list
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: cursor
          in: query
          required: false
          description: Opaque token returned as next_cursor in the previous response. Pass it unchanged; omit it for the first page.
          schema:
            type: string
      responses:
        '200':
          description: Products list successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        price_cents:
                          type: integer
                          minimum: 0
                          description: Product price in Ukrainian kopiykas
                        stock:
                          type: integer
                          minimum: 0
                          description: Number of units available in stock
                      required:
                        - id
                        - name
                        - price_cents
                        - stock
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque token for the next page. Null means there are no more pages.
                required:
                  - items
                  - next_cursor
        '500':
          description: An internal error occurred while retrieving products
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Resource not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '400':
          description: Invalid pagination parameters or cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    post:
      operationId: addProduct
      summary: Add product to the list
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
          description: Same key with the same body returns the same response without repeating creation. Replay returns 201 and Idempotency-Replay true; a different body returns 422. Keys are retained until process restart.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                stock:
                  type: integer
                  minimum: 0
                price_cents:
                  type: integer
                  minimum: 0
              required:
                - name
                - price_cents
                - stock
      responses:
        '201':
          description: Successfully created product or replayed creation
          headers:
            Idempotency-Replay:
              description: True when a stored response is replayed
              schema:
                type: boolean
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  stock:
                    type: integer
                    minimum: 0
                  price_cents:
                    type: integer
                    minimum: 0
                required:
                  - id
                  - name
                  - price_cents
                  - stock
        '400':
          description: Product info is incorrect
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unathorized user cannot create
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: This operation is forbidden for current user
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: An internal error occurred while creating product
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Idempotency-Key was already used with a different body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /products/{id}:
    get:
      operationId: getProductById
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            type: integer
      responses:
        '200':
          description: Product successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  price_cents:
                    type: integer
                    minimum: 0
                    description: Product price in Ukrainian kopiykas
                  stock:
                    type: integer
                    minimum: 0
                    description: Number of units available in stock
                required:
                  - id
                  - name
                  - price_cents
                  - stock
        '400':
          description: Invalid product ID; an integer is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Product with the specified ID was not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error or response contract violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
        - type
        - title
        - status
        - detail
        - instance